pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-monitor"
version = "1.0.0"
description = "Monitor Telegram, Twitter/X, RSS feeds, and Cloudflare Radar — push alerts to WhatsApp, Signal, Discord, Slack, or email."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "OSINT Monitor Contributors" },
]
keywords = ["osint", "monitoring", "telegram", "twitter", "rss", "alerts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "telethon>=1.36",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "feedparser>=6.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
email = ["aiosmtplib>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
osint-monitor = "src.main:cli_entry"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
